requirements.txt
bottle
pandas
xlrd
helium
python-selenium
aiohttp
pony
bs4
requests
python-dateutil
dateparser
whoosh
olefile
pendulum
bottle-cors
